--- app/src/main/java/com/syabonbubble/profilecode/DataBase/User/UserDao.kt
@@ -1,16 +1,0 @@
-package com.syabonbubble.profilecode.DataBase.User
-
-import androidx.lifecycle.LiveData
-import androidx.room.*
-
-@Dao
-interface UserDao {
-    @Query("SELECT * from user_table")
-    fun getAll(): LiveData<List<User>>
-
-    @Insert(onConflict = OnConflictStrategy.IGNORE)
-    suspend fun insert(user: User)
-
-    @Query("DELETE FROM user_table")
-    suspend fun delete()
-}
