--- app/src/main/java/com/syabonbubble/profilecode/DataBase/User/UserDatabase.kt
@@ -1,32 +1,0 @@
-package com.syabonbubble.profilecode.DataBase.User
-
-import androidx.room.Database
-import androidx.room.Room
-import androidx.room.RoomDatabase
-import android.content.Context
-
-@Database(entities = [User::class], version = 1)
-abstract class UserDatabase : RoomDatabase() {
-
-    abstract fun userDao(): UserDao
-
-    companion object {
-        @Volatile
-        private var INSTANCE: UserDatabase? = null
-
-        fun getDatabase(
-            context: Context
-        ): UserDatabase {
-            return INSTANCE ?: synchronized(this) {
-                val instance = Room.databaseBuilder(
-                    context.applicationContext,
-                    UserDatabase::class.java,
-                    "user_database"
-                ).fallbackToDestructiveMigration().build()
-                INSTANCE = instance
-
-                instance
-            }
-        }
-    }
-}
